jinahub/indexers/SimpleIndexer/requirements.txt
jina-commons @ git+https://github.com/jina-ai/jina-commons.git#egg=jina-commons
